back-end/requirements.txt
django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
celery==5.3.1
Pillow==9.5.0
djangorestframework-simplejwt==5.3.0
google-cloud-storage==2.11.0